feat: dependency inversion
