+++ packages/core-browser/src/common/common.contribution.ts
@@ -9,7 +9,6 @@ import {
   Event,
   replaceLocalizePlaceholder,
 } from '@opensumi/ide-core-common';
-import { TERMINAL_COMMANDS } from '@opensumi/ide-terminal-next';
 
 import { IContextKeyService, IContextKey } from '../context-key';
 import { corePreferenceSchema } from '../core-preferences';
@@ -19,7 +18,7 @@ import { MenuContribution, IMenuRegistry, MenuId } from '../menu/next';
 import { PreferenceContribution } from '../preferences';
 import { AppConfig } from '../react-providers/config-provider';
 
-import { FILE_COMMANDS, COMMON_COMMANDS, EDITOR_COMMANDS } from './common.command';
+import { FILE_COMMANDS, COMMON_COMMANDS, EDITOR_COMMANDS, TERMINAL_COMMANDS } from './common.command';
 import { ClientAppContribution } from './common.define';
 
 export const inputFocusedContextKey = 'inputFocus';
