+++ packages/terminal-next/src/common/commands.ts
@@ -1,118 +1,1 @@
-export namespace TERMINAL_COMMANDS {
-  const CATEGORY = '%terminal.name%';
-
-  export const SPLIT = {
-    id: 'terminal.split',
-    label: '%terminal.split%',
-    category: CATEGORY,
-  };
-
-  export const ADD = {
-    id: 'terminal.add',
-    label: '%terminal.new%',
-    category: CATEGORY,
-  };
-
-  export const REMOVE = {
-    id: 'terminal.remove',
-    label: '%terminal.remove%',
-    category: CATEGORY,
-  };
-
-  export const CLEAR = {
-    id: 'terminal.clear',
-    label: '%terminal.clear%',
-    category: CATEGORY,
-  };
-
-  export const OPEN_SEARCH = {
-    id: 'terminal.search',
-    label: '%terminal.search%',
-    category: CATEGORY,
-  };
-
-  export const SEARCH_NEXT = {
-    id: 'terminal.search.next',
-    label: '%terminal.search.next%',
-    category: CATEGORY,
-  };
-
-  export const OPEN_WITH_PATH = {
-    id: 'terminal.openWithPath',
-    category: CATEGORY,
-  };
-
-  export const CLEAR_CONTENT = {
-    id: 'terminal.clearContent',
-    label: '%terminal.menu.clearCurrentContent%',
-    category: CATEGORY,
-  };
-
-  export const CLEAR_ALL_CONTENT = {
-    id: 'terminal.clearAllContent',
-    label: '%terminal.menu.clearAllContents%',
-  };
-
-  export const TAB_RENAME = {
-    id: 'terminal.tabRename',
-  };
-
-  export const SELECT_ALL_CONTENT = {
-    id: 'terminal.selectAllContent',
-    label: '%terminal.menu.selectAllContent%',
-  };
-
-  export const MORE_SETTINGS = {
-    id: 'terminal.moreSettings',
-    label: '%terminal.menu.moreSettings%',
-  };
-
-  export const SELECT_TYPE = {
-    id: 'terminal.selectType',
-  };
-
-  export const SELECT_ZSH = {
-    id: 'terminal.selectTypeZsh',
-    label: 'zsh',
-  };
-
-  export const SELECT_BASH = {
-    id: 'terminal.selectTypeBash',
-    label: 'bash',
-  };
-
-  export const SELECT_SH = {
-    id: 'terminal.selectTypeSh',
-    label: 'sh',
-  };
-
-  export const SELECT_CMD = {
-    id: 'terminal.selectTypeCMD',
-    label: 'cmd',
-  };
-
-  export const SELECT_POWERSHELL = {
-    id: 'terminal.selectTypePowerShell',
-    label: 'powershell',
-  };
-
-  export const COPY = {
-    id: 'terminal.copy',
-    label: '%edit.copy%',
-  };
-
-  export const PASTE = {
-    id: 'terminal.paste',
-    label: '%edit.paste%',
-  };
-
-  export const SELECT_ALL = {
-    id: 'terminal.selectAll',
-    label: '%terminal.menu.selectCurrentContent%',
-    category: CATEGORY,
-  };
-
-  export const RE_LAUNCH = {
-    id: 'terminal.relaunch',
-  };
-}
+export { TERMINAL_COMMANDS } from '@opensumi/ide-core-browser/lib/common/common.command';
